章节、单元学习记录接口测试版 - 1.0

+++ family-cc/src/main/java/com/family/cc/controller/CcUnitController.java
@@ -39,7 +39,7 @@ public class CcUnitController extends BaseController {
      * @return
      */
     @GetMapping("{unitId}")
-    public AjaxResult getUnit(@PathVariable("unitId") Integer id) {
+    public AjaxResult getUnit(@PathVariable("unitId") Long id) {
         return ccUnitService.getUnit(id);
     }
 }
